+++ test/hermes_test/test_utils.py
@@ -5,6 +5,7 @@
 
 import toml
 from pathlib import Path
+from argparse import Namespace
 
 pyproject = toml.load(Path(__file__).parent.parent.parent / "pyproject.toml")
 expected_name = pyproject["project"]["name"]
@@ -22,3 +23,8 @@ def test_hermes_user_agent():
     assert (
         hermes_user_agent == f"{expected_name}/{expected_version} ({expected_homepage})"
     )
+
+def test_mask_values_options():
+    from hermes.utils import mask_options_values
+    ns = Namespace(foo="bar", options=[("foo", "bar"), ("bar", "foo")])
+    assert mask_options_values(ns) == Namespace(foo="bar", options=[("foo", "***REDACTED***"), ("bar", "***REDACTED***")])